Move CORS handling from middleware to preview route handler
- Removed middleware.ts that handled CORS for /preview route
- Added CORS headers directly in preview route handler
- Added OPTIONS handler for preflight requests
- Included CORS headers in all response types (success and errors)

This allows external clients like Architect to make cross-origin requests to the preview endpoint.
